project/server/requirements.txt
flask==2.2.2
flask-sqlalchemy==2.5.1
flask-restful==0.3.9
flask-Migrate==3.1.0
flask-wtf==1.1.1
email_validator==1.2.1
bcrypt==4.0.1
alembic==1.9.4
flask-security-too
flask-security-too[fsqla,common]
